LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/cookie-indices.https_include=changedWithRedirect3-expected.txt


FAIL If the redirect needs to be rerequested and goes elsewhere, we cannot can use the prefetched final response. assert_equals: expected (undefined) undefined but got (string) "prefetch"
